missions as behavior
